+++ PKHeX.Core/Legality/Encounters/Generator/ByGeneration/Lump/EncounterGenerator7X.cs
@@ -8,8 +8,8 @@ public sealed class EncounterGenerator7X : IEncounterGenerator
 
     public IEnumerable<IEncounterable> GetPossible(PKM pk, EvoCriteria[] chain, GameVersion game, EncounterTypeGroup groups) => pk.Version switch
     {
-        (int)GameVersion.GO => EncounterGenerator7GO.Instance.GetPossible(pk, chain, game, groups),
-        > (int)GameVersion.GO => EncounterGenerator7GG.Instance.GetPossible(pk, chain, game, groups),
+        GameVersion.GO => EncounterGenerator7GO.Instance.GetPossible(pk, chain, game, groups),
+        > GameVersion.GO => EncounterGenerator7GG.Instance.GetPossible(pk, chain, game, groups),
         _ => EncounterGenerator7.Instance.GetPossible(pk, chain, game, groups),
     };
 
@@ -21,8 +21,8 @@ public sealed class EncounterGenerator7X : IEncounterGenerator
 
     public IEnumerable<IEncounterable> GetEncounters(PKM pk, EvoCriteria[] chain, LegalInfo info) => pk.Version switch
     {
-        (int)GameVersion.GO => EncounterGenerator7GO.Instance.GetEncounters(pk, chain, info),
-        > (int)GameVersion.GO => EncounterGenerator7GG.Instance.GetEncounters(pk, chain, info),
+        GameVersion.GO => EncounterGenerator7GO.Instance.GetEncounters(pk, chain, info),
+        > GameVersion.GO => EncounterGenerator7GG.Instance.GetEncounters(pk, chain, info),
         _ => EncounterGenerator7.Instance.GetEncounters(pk, chain, info),
     };
 }
