+++ PKHeX.Core/Legality/Evolutions/EvolutionGroup/EvolutionGroup3.cs
@@ -6,7 +6,7 @@ public sealed class EvolutionGroup3 : IEvolutionGroup
 {
     public static readonly EvolutionGroup3 Instance = new();
     private static readonly EvolutionTree Tree = EvolutionTree.Evolves3;
-    private const int Generation = 3;
+    private const byte Generation = 3;
     private static PersonalTable3 Personal => PersonalTable.E;
     private static EvolutionRuleTweak Tweak => EvolutionRuleTweak.Default;
