+++ PKHeX.Core/Legality/Evolutions/EvolutionGroup/EvolutionGroup5.cs
@@ -6,7 +6,7 @@ public sealed class EvolutionGroup5 : IEvolutionGroup
 {
     public static readonly EvolutionGroup5 Instance = new();
     private static readonly EvolutionTree Tree = EvolutionTree.Evolves5;
-    private const int Generation = 5;
+    private const byte Generation = 5;
     private static PersonalTable5B2W2 Personal => PersonalTable.B2W2;
     private static EvolutionRuleTweak Tweak => EvolutionRuleTweak.Default;
